Don't rely on Referer header for event page
Turns out Firefox/Safari ignore the refererPolicy setting and only ever
send the origin anyway, sigh. So instead, we directly set the page in
the body of the event and use that on the server.

Sorting this out also reminded me that knowing the actual referer would
be a useful/interesting piece of data, so on the server we repurpose the
referer field for the actual referer and add a new page field for the
page. This uses the DB migration logic to set-up/migrate, etc.